benchmarks/requirements.txt
# pyroboplan -> uses differential pinocchio ik
# pin-pink -> uses pinocchio ik
# ikpy
ManipulaPy
# pinocchio python
genesis-world
roboticstoolbox-python
